librarian/requirements.txt
jinja2
requests
pyyaml
ghapi
